Tricks to accelerate rendering of tracks.
On MacOSX, drawing lines has become incredibly slow. However, I could
not date when the problem appeared, nor what causes it. I am note sure
if it is even serious, because there are not than many echoes of this on
the web (however, look at this:
http://stackoverflow.com/questions/30861201/why-is-java-awt-graphics-drawline-exceptionally-slow
in particular the benchmarks).

This commit ships two attempts to workaround this problem, but does not
address its core:
- draw thin line (thickness 1f instead of 2f) for normal tracks.
- only draw tracks in the rendered rectangle.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>