Fix a design fault causing massive slowdown when manually adding a spot.
Manual ("pointwise") editing of the model is a crucial feature of
TrackMate.

However it was noted than adding a spot ('A' key with the default view)
is excruciatingly slow for models with a large number of spots.

It turns out that the logic in charge of updating the model numerical
features after a spot addition was badly designed. Indeed, adding a spot
triggered an event that was causing the features to be recalculated for
ALL spots in the model, even if we added just one. This can take a very
long time.

This fix does not change the API. Instead, we check if the spot analyzer
we have to deal with is an IndependentSpotFeatureAnalyzer. Then we cast
it onto this class, and use to process only the spots that have been
changed.

This results in a massive acceleration.

Noticed by plently of people, including @etarena, Anna Gilles and
myself.

Fixes #62.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>