Change the LoG kernel normalization.
We deal with images that have a physical calibration, and this
calibration might be different across X, Y and Z.
With this commit, the quality value returned by the LoG detector is such
that:
- It has a maximal value for spots that have the size this kernel is tuned
to (for equal spot intensity).
- The quality will be of the same order of magnitude than the raw spot (if it
has the right size).
- If the image has its calibration changed by a constant factor, one will
retrieve the same quality value value than before scaling.  However, I (JYT)
could not derive the exact formula if the image is scaled differently across X,
Y and Z.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>