The 3D Viewer does not try to keep in sync with model edits.
It simply did not work, but was claiming it was.
This boils down to my (JYT) inability to modify an ImageJ3D Content
on the fly with elegance, or to add a new Content to a time sequence
in the ImageJ3D.

So we officially give up. The 3D Viewer just shows a snapshot of the model
at a given time.

As said to the users:

Also note that a 3D view is not kept in sync with manual editing of the model.
If you manually edit the model (add, remove, move or modify a spot; delete,
cut, merge a track, etc...) this view will not show the modifications.
It contains an immutable snapshot of the model taken at the time when
it was launched.