Implement step-wise time browsing, 'a la MaMuT'.
The F & G key can be used to browse in time, using discrete steps.
The size of these steps are defined in the SpotEditRoi control
panel. The jumps are made to fall always on the same frame (e.g.
with a period of 5 frames, you always get frames 1, 6, 11, etc...
but never frames 2, 7, 10 etc..).
This facilitates manual annotation of rich, large movies.