Release v3.1.0.
Improvements:

* Add an action that export all visible spots statistics, regardless of whether
they belong in a track or not.

* New display mode: show only selection.

This track display mode only shows the content of the current selection, for
spots and edges.  It can be accessed in the GUI panel 'Configure views' with a
special track display mode called 'Selection only'.

Careful: the content of the selection model can still be edited in this mode.
Of course, the track display mode is ignored by TrackScheme, which then can be
used to modify the selection in this mode.

Changes:

* We really do not relaunch TrackScheme automatically when loading a
TrackMate file.