Groovy script used to make TrackMate macro recordable.
Contributed by @imagejan during a NEUBIAS course.
Quoting from Jan:
"The macro language is too limited to work with such awesome
things as TrackMate, but that you can do everything with a more
powerful scripting language. So when using a 64-line script to
call it, it actually is macro recordable."