Release v3.6.1.
Change how we write track and edge: enforce proper orientation of edges.
If in some cases an edge has the wrong orientation, we will still
save it to XML with the proper orientation (forward in time). This
is to avoid misfits with 3rd party softwares that load data from the
XML expecting the edge orientation to be constant.