Set the nameID for DefaultNameGenerator on open command
The nameID in DefaultNameGenerator is set to zero, resulting in the duplicated track names in the combination of the following situations.
-  The MaMuT annotation results containing the tracks with the name pattern of "^Track_[0-9]+$" are opend via MaMuT > Open MaMuT annotation.
- New tracks are generated.
This commit resolves the above issue by setting the nameID to the smallest integer value that is greater than any nameIDs used in the tracks loaded from the annotation file.