Fix important TrackMate bug in linking.
Our Jonker-Volgenant algorithm implementation does not work with
costs == 0. This results in track breaks when trying to link 2 spots
that have exactly the same positions.

I probably did not understand the meaning of "positive costs"
in the paper. :(

Mainly this happened when tracking spots that were detected with the
LoG detector without the sub-pixel accuracy. The fact that we used
this sub-pixel accuracy step by default since the beginning of
TrackMate prevented us from seeing this bug in front of my face.

The fix is to prevent costs == 0 in linking with the LAPVJ by substiting
Double.MIN_NORMAL when we should have 0.