An action to export tracking results to a label image.
This action creates a label image from the tracking results. "

A new 16-bit image is generated, of same dimension and size that
of the input image. The label image has one channel, with
black baground (0 value)
everywhere, except where there are spots. Each spot is painted with
a uniform integer value equal to the trackID it belongs to.
Spots that do not belong to tracks are painted with a unique integer
larger than the last trackID in the dataset.

Only visible spots are painted.