Addd a flag to forbid running a module concurrently.
This might be the case for instance if you have a spot detector
that you want to run strictly one frame after another, to save
resources or because the multi-threading of the detector cannot
be controlled. In that case this flag should be set to true.

If false (the default), the multi-threading will be decided by
TrackMate.

So far we use this only for the detectors where it can be important.