Support 2D Roi in TrackMate.
This time we story the polygon arrays, in physical coordinates, relative
to the spot center, as a new field in the Spot class.

This attempt is much better than the previous one:

1. Code is more compact. There is very little special logic to harness
the ROIs.
2. There is no import of ij.gui.Roi.
3. Which turns out takes much less RAM.
4. Spot painted with ROI abide color, etc... And they can be
interactively dragged like normal spots.

It works simply and I like it a lot.
What is left:
- Find and fix bugs.
- De/serialize (will be tough.
- Think about how we compute features now. Shall we abide to the ROI?
- A display option to switch between ROI and circles?