Rework the detector hierarchy to make it possible to have global detectors.
We want to make it possible to implement detectors that do not operate
frame by frame, but work on all the time-points at once. This commit adds
to the existing detector hierarchy to make it possible.

Not tested in depth yet. If I come back to read this commit in anger,
this is probably the reason. But it's late.