A detector that creates spots by thresholding a grayscale image.
More precisely, all the pixels in the designated channel that have
a value strictly larger than the threshold are considered as part
of the foreground, and used to build connected regions.
In 2D, spots are created with the (possbily simplified) contour
of this region. In 3D, a spot is created for each region in its
center with a volume equal to the region volume.