A detector that creates spots from a black and white mask.
More precisely, all the pixels in the designated channel that have
a value strictly larger than 0 are considered as part of the foreground,
and used to build connected regions. In 2D, spots are created with
the (possibly simplified) contour of this region. In 3D, a spot is
created for each region in its center with a volume equal to the
region volume.

This just the Threshold-detector, with one parameter less: the intensity
threshold is set to 0.