A class that can generate configurable feature coloring selector.
Can create panels, managed by the same DisplaySettings object (and
properly wired to it), that let users select a feature, a min & max
display range (with an auto min max button), and a colormap.
This sets fields in the display settings, that can be used to
generate a feature color generator.