Fix bugs in the spot-edit tool
It was possible to create backward edges.
Normally we want to enfore edges to go forward in time. It is important
for external programs that read the TrackMate XML file. This was the
probably source of the issues that have been reported orally some time.

Also fix a bug, introduced when we added spot ROIs, that prevented the
radius of spots to be changed when there was no ROIs.