Serialize DisplaySettings to XML via Gson.
Stefan & friends if you read this don't be angry at me. It's just that
Gon is so convenient and XML is not at all. Look at how shorter are
my classes, and the resulting XML is ok looking too.
Plus I don't have to worry about adding new fields in this display
settings thingie.