Intercalate in the spot analyzer factory hierarchy.
Later we want to introduce analyzers that can only operate
properly on spots that have ROIs. We may want to treat them
separately in the GUI: it makes no sense to add such analyzers
for detectors that only return spots as points (like the
canonical LoG detector).
So we add a mother interface SpotAnalyzerFactoryBase that has
two daughter interfaces:
- SpotAnalyzerFactory for classical analyzers.
- SpotMorphologyAnalyzerFactory for analyzers that can exploit
ROIs.
They are otherwise identical. Clients have to choose one
or the other and the decision to add them to analysis will
for the application.