When loading: add all analyzers.
Backward compatibility: for now we add all the analyzers we know of
at runtime. This kind of kills the need to read the analyzers list
specified in the TrackMate file, but this gives a fully working
application with the file loaded, even with TrackMate <7.0.0.,
and that can be resaved without issues.