Rework the feature grapher.
We now have specialized JFreeChart datasets that wrap a
TrackMate model, instead of copying the data from the latter
to the former. Then we can exploit a common hierarchy to
shrink the number of lines.
WIP: if TrackMate loads an unknown feature, the grapher
breaks because it does not know what is the dimension of
the unknown feature.