When loading a TM file, declares the runtime analyzers.
This is required mainly for backward compatibility.
We might want to load TM files that declare features that are not
known at runtime. This is fine and handled properly. The
feature values are loaded and declared with dimension etc.

But we also want to declare in the model the feature analyzers that are
discovered at runtime (coming from the TM program), but not declared
in the TM file. These analyzers are already present in the Settings
objects, and the feature associated will be computed during plugin use.
So it is a good idea for the Model to know there name, dimension etc
even if there is no value associated yet.