Fix bug with manual editing in TrackMate.
Since TrackMate v7 there I have found the following bug:

Sometimes the listeners of the spot editor tool get added to the
target image canvas TWICE. The reason is unclear to me yet, because
it happens only sometimes. But like very often, more often than not.
I would say about 80% of the time on my mac.

This causes an unspeakable mess where all events are
triggered twice for e.g. a single click. For instance you cannot
shift-click on a spot to add it to the selection, because the event
is fired TWICE, which results in the spot being de-selected
immediately after being selected.

But the double registration seems to happen randomly. Sometimes the
listeners are added only once, *sometimes* (more often) twice.

To work around this mess, we overload the registerTool(ImageCanvas)
method and skip the registration if we find that the mouse listener
has already been added to the canvas. It fixes the issue, regardless
of the occurrence of the double call to this method or not.