Add a copy method to the detector and tracker factory.
This will break the external detectors we added in v7, so they will
need an update too.
This is required to properly copy a Settings object without any
undesirable side effect. All in all, the issue we are trying to address
is a design fault done early in TrackMate, where we allowed a tracker
and a detector factory to have instance fields.... What a mistake...