An action to export to a MotilityLab spreadsheet.
MotilityLab is a R application that performs track analysis. One instance
is running into a webiste: http://www.motilitylab.net/startpage.php

The way to import data into the website is to copy-paste a spreadsheet
with sepcified columns in a given order and format.

This action displays the current tracking results in such a spreadsheet,
with the columns pre-formatted, so that it can directly copy-pasted in
the website without any extra manipulation.
- In the TrackMate table: Ctrl+A, Ctrl+C
- In the MotilityLab import page: Ctrl+V