Check for possible consistency errors.
It is possible that 2 spots are so close than the mask of one completely
erase the other. This will trigger an error in the CTC metrics
calculation so we have to protect ourselves from it. This solution
consists in check whether that is the case, and if yes, removing the
erased spot from the smallest track.

This fix is approximate: the calculus for complete overlap assume the
spots are spherical with a fudge factor.