Add a safeguard when saving against badly defined features.
It is possible to reach the save point with features that have been
improperly defined. For instance, a feature might not have declared
its 'isInt' flag. This happens for instance when analyzing an image
with two channels, but forgetting to call the #setNChannels( 2 )
method of the SpotIntensityMultiCAnalyzerFactory.
This is undesirable but it should not prevent users from saving
the data. This commit adds a safeguard that prevents TrackMate from
crashing with such features when saving.