Make some fields in Settings private or remove them.
Recent changes in how we handle Rois in TrackMate have generated
confusions between the interval we pass to detectors and the Roi
we use to filter detection. This commit reconcile anf fuse both
notions by removing one of them :/

The Roi in settings is now a private field and has a getter and
a setter. The xstart, xend, ystart and yend field have been removed
and a getter for each returns meaningful value based on the roi
field content (null is ok).

This might break some scripts.