Use the ROIs in the RoiManager as fields for the Hessian detector.
If the RoiManager is visible and has rois in it, they will be used
as regions to be analyzed separately with the Hessian detector.
This is useful when a single image comtains different regions of
very different brightness. Soits found in bright regions will mask
the ones in faint regions even if they are relevant.
By analyzing these regions separately and using a normalized quality,
one can retrieve them all.
This also optimize computation time.