New action that computes the distance from spot to closest ROI boundary.
Computes the distance from each spot to the closest ROI in the current
frame. The distances are stored in a new spot feature and the distances
are reported using physical distance units.

The ROIs are taken from the Roi Manager window currently opened. All the
ROIs from the current time-point are considered, so the frame position
of the ROIs must be properly set. If the frame position of a ROI is not
set, it will be considered as being present for all time-points.
The Z position of the ROIs is not taken into account. That is: the
distances are calculated as if all the ROIs would be in the Z plane
of the spot.