Fix SpotIntensityMultiCAnalyzer for small spots.
When spots are small (in pixels) there is a chance that the estimate
of the number of pixels to iterate to become wrong by 1 pixel. This
causes a crash.
Now we do not assume anything about the size of the spot iterable
and use the SciJava DoubleArray facility. We also implements method
to compute mean std and sum in such data structure.