A utility that handles previewing detection and shows a quality histogram.
The class DetectionPreview handles all what is required to preview
detection on a single frame. It offers a panel that can be included
in the configuration panels, that has the button, the preview log, and
a nice quality histogram that can be interacted with
- left clik to set the threshold
- right click for auto-threshold (Otsu)
- key supported (left, right, typing a numerical value)

It behaves a bit differently depending on whether the detector
has a TRESHOLD parameter on quality: If yes, the quality histogram
include ALL spots even those below the quality value, but a marker
shows where the threshold is. If not, it behaves classically.