Make the gap-closing methods configurable with a list of double params.
Each implementation must return a list of params (embedded in a simple
value class). It is used in the GUI to generate a list of slider that
enables the user to set a value for this parameters, then used in the
gap-closing process.

This is currently used in the 'gap-closing by detection' method by
allowing to set the factor for the search radius.