Fix gap-closing by detection sometimes splitting tracks in pieces.
Noticed by @gletort.

Also changed how we compute the position around which to perform
detection.
Before we simply relied on the interpolated positions in a gap. For long
gaps this position could become increasingly inadequate when the cell
is moving back and forth a lot.
Now what we do is that after we have found a new spot, we recompute
the interpolated positions for next frames from this new spot.