A simple data structure to store a 3D mesh in a Spot object.
This class is the 3D counterpart of SpotRoi. It stores the object
shape as a mesh, and has (for now) a few methods to facilitate
painting it and creating it.
The mesh are stored with coordinates relative to the spot center
(mesh center is at 0,0,0). The same for the bounding box. The
mesh coordinates are expected to be in physical coordinates, not
pixel coordinates.