Paint 3D meshes in the HyperStackViewer.
We simply reslice the mesh at the Z slice currently displayed,
and paint the intersection as a collection of segments.

I also took the opportunity to refactor a bit the SpotOverlay.

Right now this works but is not optimal:
1/ There are weird stuff happening at the *top* of the mesh: it's
like we miss some part of it.
2/ The slice routine generates a list of disconnected segments. It
does not show when we paint them, but maybe would be nice to
reconstruct the collection of contours resulting from the intersection
of a mesh with a plane.
3/ We could optimize the slice() routinemaybe  by having an index
that sorts triangles by their minZ value, and another index that
sorts them by their maxZ value. This way we could quickly retrieve
the triangles to sort by two binary-search and one set intersection.