Serialization and deserialization of spot meshes to file.
The meshes are saved in a ZIP file next to the TrackMate XML file.
If the TrackMate file is called: trackmate-file.xml, then the file
that stores the meshes for the model are in a zip file called
trackmate-file.xml.meshes

This zip file just contains the meshes as PLY files, named with the
spots ID. If a spot has an ID equal to 1234, then the PLY file that
stores its mesh is named 1234.ply in the zip file. If the zip file
does not contain such a file then it means that this spot does not
have a mesh. If there is not .meshes file, it means that no spot
have a mesh.

Possible improvement: read all meshes at once, instead of opening and
closing the zip file for every spot.