Rework spot mesh iteration.
Use the contours we got from the ZSlicer. Perform ray cast along
the X axis. Count how many intersection we crosses to determine
whether we are inside or outside.
Works when the slice has several disjoint contours and when some
contours are surrounding the exterior of the mesh (holes inside
the slice).

Relatively optimized:
- Once per spot: get all the Z slices. Involves iterating once through
all the triangles, sorting 2 arrays and a few binary search.
- Once per X line: compute intersections of a ray along the X axis
with all the contours.
- Once per pixel: binary search against intersection to know how
many of them we crossed.