Create a proper mask from the input before using the mask detector.
A small routine builds a view of the input (that should be a mask)
where all pixels greater than 0 are set to 1 and to 0 otherwise.
In the case of meshes, this allows using the marching cube algorithm
on real type, interpolating at 0.5 between 0 and 1, and having
smoother meshes than when using the marching cube on boolean types.