Precompute and store a cache of Z-slices for meshes.
We don't want to recompute them every time we have to paint the
spot-meshes, or iterate through their pixel. And we have an
optimized version of the ZSlicer that computes Z-slices in batch
with several Zs. This makes the display and iteration in the
TrackMate app much much faster for large meshes.

The difficulty is that we need to specify a scale in XY and in
Z (to simplify the  contours, and compute the Z positions of
the intersections with the image stack slices). But we don't
want to store these scales - which are the pixel size typically -
in the model, that should be independent of a reference to an image.

The solution in this commit consists in receiving these
scales when we require the precomputed Z-slices: the cache is
then recomputed if needed, and we have these scales in the code
when we require the Z-slices