A UI tool to build a selection of feature analyzers and save it to disk.
Later we will make the GUI sessions use the saved selection, so that
users can choose what they want to compute or not.
Might be handy for some features that are long to compute.
Contrast and SNR I am looking at you. And actually you are deselected
by default now.