Rework the 2D and 3D detection utils.
We have now 3D meshes that we get via the marching cubes. This
algorithm behaves slightly differently for grayscale + threshold
and mask images, so we have to implement that in TrackMate.

First: split the shape-related methods of MaskUtils in two
utility classes SpotRoiUtils and SpotMeshUtils so as to avoid
having a single gigantic class.

With grayscale thresholded image, the marching cube algorithm can
return nice, smooth, interpolated meshes, which is what we want
in a biological context. So there is now in SpotMeshUtils a method
that exploits this in the following manner:

The grayscale marching-cube algorithm is used to create one big mesh
from the source image. It is then split in connected-components to
create single spot objects. However, to deal with possible holes
in objects, meshes are possibly re-merged based on full inclusion
of their bounding-box. For instance, a hollow sphere would be
represented by two connected-components, yielding two meshes.
But because the small one isincluded in the big one, they are
merged in this method so that the hole is properly included in the
shape representation.

When dealing with masks as inputs, we want to generate a mesh that
can retrieve the exact pixel content of the mask when iterated. So
a separate method convert the mask into a label image, and each
of its connected-component is treated separately to generate a mesh.
We will use the grayscale marching-cube algorithm, using a threshold
value of 0.5 on the bit-masks resulting from connected-component
analysis of the label image generated from the mask.