Convenience class to run 2D+Z segmentation algos in 3D.
A SpotDetector for 3D images that work by running a spot
segmentation algorithm on 2D slices, and merging results
using a tracker. This yield a label image that is then
converted to 3D meshes using the LabelImageDetector.

This is a convenience class, made to be used in
specialized SpotDetectorFactory with specific choices
of detector and merging strategy.