WIP: A new viewer for TrackMate based on the BVV.
Adapting the demo from Tobias. This currently requires the 'mesh' branch
in the bvv repo to work. Right now it can display the 3D image (possibly
multi-channel) over time, along with the meshes in spots (if there is
no mesh, nothing is shown).
There is still so much to do but this is cool!