A LabKit launcher made to edit spots in the current frame.
This launcher catches what time-point is currently diplayed,
and wrap all information needed for LabKit to edit the spot
it contains, as labels.

While the user edits the labels in LabKit, the TrackMate UI is
disabled.

Once the user closes the LabKit window, the modifications they
made are inspected. This class compares the new labels with the
previous ones, and can determine whether a spot has
been added, removed or modified. In the last case it updates the
model with the modified spots, and reintroduces it in the tracks
as it should.

All features of modified spots and their edges are recomputed.

If a label has several connected components, they are added as
separate spots. The one closest to the original spot is reintroduced
in the tracks.

The label names are important: they are used to retrieve the original
spot id and the original spot shape for comparison.
If the user modifies a label, it will be perceived as a
new spot instead of a modified one.

Tested in 2D so far.