Rework and improve the import method of the spot editor.
- We don't depend on labels anymore, but directly operate and compare
the index images (before modification and after). Because the index
is directly related to the spot ID, we can get a match from previous
spot to novel spot in an easy manner.

- The spots from the edited version are created directly from the
novel index image, using something adapted from the label image
detector code, so again, just one pass. We use the fact that we can
provide it with a 'quality' image, and read the index of the label
image 'under' the spot and write it into its quality value.
This way we can retrieve the id of the matching previous
spot in an easy manner.

- The price to pay for not working with labels anymore
is that we don't have access to the label name, but that's life.

- We make only one pass over the image to collect the ids of the spots
that have been modified, instead of one pass per spot. Also, this
pass is multithreaded (thanks LoopBuilder).

- I have also learned that I should not use weakListeners() if I
am doing something with threads inside the listener. Using listeners()
instead works, but I do not know why the other one does not.
Probably something arcane with Java WeakReferences being collected.

- As a result of all this the performance is much better than before
and the 'return to TrackMate' should happen without the user noticing
the process.