Make sure the user selection of analyzers are added in priority order.
This is crucial, as some analyzers depend on others to have been
computed. This is implemented following the priority of the analyzer
factories, which is known by the provider. The provider returns a list
of keys, in priority order, so we must make sure to reproduce this
order with the user selection.
Hence the gymnastic in the configure(Settings) method.