Implement smoothing of images before segmentation.
Changes in the core methods in MaskUtils, SpotMeshUtils and
SpotRoiUtils:

- They take a smoothingScale double parameter.

- If negative, it is ignored, and the segmentation (by mask, threshold
and label image) processes directly on the input image.

- Otherwise, the inputs are filtered by a Gaussian filter with sigmas
derived from the specified scale. In 2D and 3D. Even for the bitmasks
of the label image. This yields much smoother meshes and contour,
of a smoothness controlled by the scale.

- This is different from the 'simplify' flag, that prunes the contour
and the mesh, but still 'sticks' to the shape in the input without
filtering.