Keep imagej-legacy out of the dependency tree
This works around a test failure in TrackMatePluginTest caused by the
original ImageJ classes being loaded too soon. There are other ways of
working around this problem, such as creating a SciJava context more
eagerly, but it's a distraction from the business of testing TrackMate,
and anyway we don't need imagej-legacy on the classpath.

The only reason labkit-ui has imagej-legacy as a dependency is for one
line of code, which can be refactored to avoid it, so hopefully the
exclusion will be able to disappear later after labkit-ui is updated.