Mecanism to allow selecting among several arguments in a CLI.
For instance, in cellpose we have the --pretrained_model arg that
can accept either a string from a pretrained model list, or a path
to a custom model. This corresponds to two different types of
Argument<?> : a list and a path.

To make it possible to choose between one OR the other in a GUI,
there is a SelectableArguments that lets the user specifies that
one argument can be active OR another one, but not bother at the same
time.

	// State that we can set one or the other.
	cli.addSelectableArguments()
			.add( ptm )
			.add( cm )
			.select( ptm );

This is supported in the GUI (but what a mess) and the args that are
in a selectable group appear with a radio button. Selecting one
properly sets the CLI.

The arguments in a selectable groups do not have to be following
in the arg list, but it is prettier that way.