Add translator to argument value in CLI tools.
Sometimes we need to change the output of an argument so that it
'looks good' in TrackMate GUI but still translates into something
the executable can understand. For instance in cellpose we want
to display for the optional channel 2:
- 0 (don't use)
- 1 (red)
- 2 (green)
- 3 (blue)
into the accepted values by the CLI:
- 0
- 1
- 2
- 3
This mechanism offers the possibility to do so.