The CLI UI builder and TrackMate map builder accept extra parameters.
Typically, a TrackMate detector or tracker has more parameters
than the CLI of the tool it will call. For instance to simplify
contours after a detection step.
This commit allows passing these extra parameters to the UI builder
and to the map serializer so that a TrackMate module that calls a
command line tool can be handled by the CLI framework.