Rework the extra parameters of the CLI config.
Do not use Elements. Instead just gives the possibity of adding
extra arguments, making them visible in the GUI, but not used in
the command. For this, add a method arg.inCLI(boolean) that specified
whether this arg is to be used in the CLI.
This allows using the same system to generate a command line, a GUI,
and manage all of this via TrackMate settings.

Also add javadocs and fix some typos.