Sub-class for CLIConfigurator that specializes in conda Python tools.
For many tools we want to integrate are Python tools that can be
run from a specific conda environment. If the implementing CLI
config knows already what is the Python command, the user should
just have to specify in what conda env it is installed.
It is more convenient than to browse to a Python executable and
less prone to error.