Harness MacOS with conda CLI tool.
We CANNOT change the conda environment with the Java ProcessBuilder.
At least after me trying very hard. So we have to defer on this
platform to calling the Python tool as a module, from the python
executable. It is very sad. And it limits what we can run this
way to tools that works as a module ('python -m blablabla').
I see no other way yey.

To do so:

- Make a special command for the arg that allows selecting a conda
environment: CondaEnvironmentCommand. Use it in the GUI and in the
command builder.

- Its translator turns it into either a 'conda activate' thing on
Windows, or in a 'python -m blablabla' on Mac.