Fix bug in LabKit editor: avoid spot ID and label conflicts.
When editing a sub-portion of the image, the spots outside the ROI
are not imported in the editor. It is possible that the user creates
a new label in LabKit that will have the same id that an existing
spot outside the ROI.
Before this fix, the spots in that case were replaced by the new ones,
creating a mess.
The fix consists in reminding what spots are imported in the LabKit
editor, then compariing to this list the new ids when reimporting
from the editor.