Fix another bug with the LabKit importer.
This time it was due to confusion between a map from spot ID to spots,
and a map from label value to spot.
I fixed it by rewriting everything in terms of label values in the
labeling. So we do not need to keep track of the spot ID anymore. The
labels could be anything. They are still equal to spot ID + 1, because
it is convenient to debug.