A new CLI configurator for executables that live in a conda env.
Mother class for CLI configurator that relies on an executable
installed in a conda environment. This happens when the Python code
we want to call does not have a Python module (that we could call
with 'python -m') or is simply not a Python program. In that case:

on Mac we resolve the env path, and append the executable name to
build an absolute path to the executable.
on Windows we simply activate the conda environment and call the
executable assuming it is on the path.