In the CLI configurator, when a key is null, let it be null.
A null key means that the argument should not appear in the
settings map when the CLI will be deserialized to the map. This is
useful e.g. for arguments of the CLI that must not be configured
by the user, nor should not be serialized to disk. Such as the path
where to save temp images.