Fix compile error with new version of imglib2.
Focuses on the rewrite of KDtree. The left and right fields of
KDTreeNode are not visible anymore. The implementation changed to be
proxy-based, so the corresponding methods, used in this commit to fix
the compile errors, are deprectaed.

We therefore shall also full rewrite the KDtree NN tracker using
the proxy approach.