Fix labkit launcher with ROIs that touch the border.
Before this commit, the labkit import "did not work".
The image appeared black, the UI was super slow etc.
Because I do not know what is really going on, and because
I already have been changed these two lines, I am keeping
them in comments, to see if I need to revert this later.