Properly manage the Labkit keymap, ...
while having a sensible, immutable default for BDV actions.

we create a specific config for BDV, so that we can use a custom
keymap selected not to interfere with the TrackMate-Labkit keymap. I
could not find a way to update the BDV keymap dynamically (the
Actions instances are private, so I could not add a listener to it).
So the only solution is to initialize the BDV window with a custom
keymap, configured rationally, and leave it as is.