Solve mess with the CLI config choice argument.
We want to have something that is displayed, and something that is
actually used in the command line. Before this commit, this was
conflated.

Now you need to do addChoice( final String choice, final String displayed )
and the logic will be based on 'choice' but the GUI will show
the 'displayed'.

This will need adapting by downstream dependencies (Trakastra, YOLO)