Generalize the CLI configurator to a general configurator.
Mainly split in two abstract classes:
- 'Configurator' for classes that specify a collection of arguments,
regardless of whether they correspond to a CLI command.
- 'CLIConfigurator' for CLI commands.

I have added the possibility to translate a value that is stored
in something different displayed in the GUI. For instance, an
argument can specify the radius, but the GUI will show the diameter.

This is the hope of using Configurator as a general way to create and
configure a TrackMate module.
We start with the hard ones: the spot detector.