A Configurator class for tools that rely on Appose.
Importantly: that rely on using mamba envuronments.yaml to download
and install tools. This abstract class defines the two abstract methods:

- getMambaYML(), to return the content of an environment.yaml file
that can download and install the tool

- getScriptTemplate(), to return a template of a Python script to
execute to run this tool. The template is supposed to contain tags
like ${--pretrained_model}, that will be replaced by the user specified
values in the configurator arguments.